Errores solucionados y clases organizadas.
