+++ src/AppForSEII2526.API/DTOs/DispostitivoParaComprarDTO/DispositivoComprarDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace AppForSEII2526.API.DTOs.PurchaseDTOs
+﻿namespace AppForSEII2526.API.DTOs.DeviceDTOs
 {
     public class DispositivoComprarDTO
     {
