+++ src/AppForSEII2526.API/Logging/RabbitMQLoggerConfiguration.cs
@@ -1,4 +1,4 @@
-namespace AppForSEII2526.API.Logging;
+/*namespace AppForSEII2526.API.Logging;
 
 public class RabbitMQLoggerConfiguration
 {
@@ -10,3 +10,4 @@ public class RabbitMQLoggerConfiguration
     public string ExchangeType { get; set; } = null!;
     public bool Durable { get; set; }
 }
+*/
