+++ src/AppForSEII2526.API/Logging/RabbitMQLoggerExtensions.cs
@@ -1,4 +1,4 @@
-using Microsoft.Extensions.Configuration;
+/*using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -15,3 +15,4 @@ public static class RabbitMQLoggerExtensions
         return builder;
     }
 }
+*/
